+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/frostserver/parser/path/PathParser.java
@@ -35,6 +35,7 @@ import de.fraunhofer.iosb.ilt.frostserver.property.NavigationPropertyMain.Naviga
 import de.fraunhofer.iosb.ilt.frostserver.property.NavigationPropertyMain.NavigationPropertyEntitySet;
 import de.fraunhofer.iosb.ilt.frostserver.property.Property;
 import de.fraunhofer.iosb.ilt.frostserver.query.Query;
+import de.fraunhofer.iosb.ilt.frostserver.query.QueryDefaults;
 import de.fraunhofer.iosb.ilt.frostserver.settings.CoreSettings;
 import de.fraunhofer.iosb.ilt.frostserver.util.StringHelper;
 import de.fraunhofer.iosb.ilt.frostserver.util.pathparser.Node.Visitor;
@@ -291,12 +292,12 @@ public class PathParser extends Visitor {
         return resourcePath;
     }
 
-    public static Query parsePathAndQuery(String serviceRootUrl, Version version, String pathAndQuery, CoreSettings settings) {
+    public static Query parsePathAndQuery(Version version, String pathAndQuery, CoreSettings settings, QueryDefaults queryDefaults) {
         int index = pathAndQuery.indexOf('?');
         String pathString = pathAndQuery.substring(0, index);
         String queryString = pathAndQuery.substring(index + 1);
-        ResourcePath path = PathParser.parsePath(settings.getModelRegistry(), serviceRootUrl, version, pathString);
-        return QueryParser.parseQuery(queryString, settings, path).validate(null, path.getMainElementType());
+        ResourcePath path = PathParser.parsePath(settings.getModelRegistry(), queryDefaults.getServiceRootUrl(), version, pathString);
+        return QueryParser.parseQuery(queryString, queryDefaults, settings.getModelRegistry(), path).validate(null, path.getMainElementType());
     }
 
 }