+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/frostserver/service/ServiceRequest.java
@@ -20,6 +20,7 @@ package de.fraunhofer.iosb.ilt.frostserver.service;
 import de.fraunhofer.iosb.ilt.frostserver.json.deserialize.JsonReader;
 import de.fraunhofer.iosb.ilt.frostserver.json.deserialize.JsonReaderDefault;
 import de.fraunhofer.iosb.ilt.frostserver.path.Version;
+import de.fraunhofer.iosb.ilt.frostserver.query.QueryDefaults;
 import de.fraunhofer.iosb.ilt.frostserver.settings.CoreSettings;
 import de.fraunhofer.iosb.ilt.frostserver.util.user.PrincipalExtended;
 import java.io.BufferedReader;
@@ -36,8 +37,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 /**
- *
- * @author jab
+ * An abstract request for the Service.
  */
 public class ServiceRequest {
 
@@ -51,38 +51,46 @@ public class ServiceRequest {
     private Version version;
     private String contentType;
     private Map<String, List<String>> parameterMap;
-    private Map<String, Object> attributeMap;
-    private PrincipalExtended userPrincipal;
+    private Map<String, Object> attributeMap = new HashMap<>();
+    private PrincipalExtended userPrincipal = PrincipalExtended.ANONYMOUS_PRINCIPAL;
     private CoreSettings coreSettings;
+    private QueryDefaults queryDefaults;
     private UpdateMode updateMode;
     private JsonReader jsonReader;
-
-    protected ServiceRequest() {
-        // empty by design.
-    }
 
     public Map<String, Object> getAttributeMap() {
         return attributeMap;
     }
 
-    public void setAttributeMap(Map<String, Object> attributeMap) {
+    public ServiceRequest setAttributeMap(Map<String, Object> attributeMap) {
         this.attributeMap = attributeMap;
+        return this;
+    }
+
+    public ServiceRequest setAttribute(String key, Object value) {
+        attributeMap.put(key, value);
+        return this;
     }
 
     public CoreSettings getCoreSettings() {
         return coreSettings;
     }
 
-    protected void setCoreSettings(CoreSettings coreSettings) {
+    public ServiceRequest setCoreSettings(CoreSettings coreSettings) {
         this.coreSettings = coreSettings;
+        if (queryDefaults == null) {
+            queryDefaults = coreSettings.getQueryDefaults();
+        }
+        return this;
     }
 
     public String getRequestType() {
         return requestType;
     }
 
-    public void setRequestType(String requestType) {
+    public ServiceRequest setRequestType(String requestType) {
         this.requestType = requestType;
+        return this;
     }
 
     /**
@@ -123,28 +131,32 @@ public class ServiceRequest {
         return contentBinary;
     }
 
-    public void setContent(InputStream content) {
+    public ServiceRequest setContent(InputStream content) {
         this.contentBinary = content;
+        return this;
     }
 
-    public void setContent(String content) {
+    public ServiceRequest setContent(String content) {
         this.contentString = content;
+        return this;
     }
 
     public String getContentType() {
         return contentType;
     }
 
-    public void setContentType(String contentType) {
+    public ServiceRequest setContentType(String contentType) {
         this.contentType = contentType;
+        return this;
     }
 
     public UpdateMode getUpdateMode() {
         return updateMode;
     }
 
-    public void setUpdateMode(UpdateMode updateMode) {
+    public ServiceRequest setUpdateMode(UpdateMode updateMode) {
         this.updateMode = updateMode;
+        return this;
     }
 
     public Map<String, List<String>> getParameterMap() {
@@ -154,8 +166,9 @@ public class ServiceRequest {
         return parameterMap;
     }
 
-    public void setParameterMap(Map<String, List<String>> parameterMap) {
+    public ServiceRequest setParameterMap(Map<String, List<String>> parameterMap) {
         this.parameterMap = parameterMap;
+        return this;
     }
 
     public String getParameter(String parameter) {
@@ -178,20 +191,31 @@ public class ServiceRequest {
         return this;
     }
 
+    public QueryDefaults getQueryDefaults() {
+        return queryDefaults;
+    }
+
+    public ServiceRequest setQueryDefaults(QueryDefaults queryDefaults) {
+        this.queryDefaults = queryDefaults;
+        return this;
+    }
+
     public String getUrlPath() {
         return urlPath;
     }
 
-    public void setUrlPath(String urlPath) {
+    public ServiceRequest setUrlPath(String urlPath) {
         this.urlPath = urlPath;
+        return this;
     }
 
     public String getUrlQuery() {
         return urlQuery;
     }
 
-    public void setUrlQuery(String urlQuery) {
+    public ServiceRequest setUrlQuery(String urlQuery) {
         this.urlQuery = urlQuery;
+        return this;
     }
 
     public String getUrl() {
@@ -201,7 +225,7 @@ public class ServiceRequest {
         return urlPath + "?" + urlQuery;
     }
 
-    public final void setUrl(String url) {
+    public final ServiceRequest setUrl(String url) {
         if (url.contains("?")) {
             this.urlPath = url.substring(0, url.lastIndexOf('?'));
             this.urlQuery = url.substring(url.indexOf('?') + 1);
@@ -209,6 +233,7 @@ public class ServiceRequest {
             this.urlPath = url;
             this.urlQuery = null;
         }
+        return this;
     }
 
     public JsonReader getJsonReader() {
@@ -227,8 +252,9 @@ public class ServiceRequest {
         return userPrincipal;
     }
 
-    public void setUserPrincipal(PrincipalExtended userPrincipal) {
+    public ServiceRequest setUserPrincipal(PrincipalExtended userPrincipal) {
         this.userPrincipal = userPrincipal;
+        return this;
     }
 
     /**
@@ -244,9 +270,11 @@ public class ServiceRequest {
      * Set the API version for this request.
      *
      * @param version the API version for this request.
+     * @return this.
      */
-    public void setVersion(Version version) {
+    public ServiceRequest setVersion(Version version) {
         this.version = version;
+        return this;
     }
 
     public static ServiceRequest getLocalRequest() {