+++ FROST-Server.SQLjooq/src/main/java/de/fraunhofer/iosb/ilt/frostserver/persistence/pgjooq/utils/validator/CheckEntityQuery.java
@@ -20,6 +20,7 @@ package de.fraunhofer.iosb.ilt.frostserver.persistence.pgjooq.utils.validator;
 import de.fraunhofer.iosb.ilt.configurable.annotations.ConfigurableField;
 import de.fraunhofer.iosb.ilt.configurable.editor.EditorString;
 import de.fraunhofer.iosb.ilt.frostserver.model.EntityType;
+import de.fraunhofer.iosb.ilt.frostserver.model.ModelRegistry;
 import de.fraunhofer.iosb.ilt.frostserver.model.core.Entity;
 import de.fraunhofer.iosb.ilt.frostserver.parser.query.QueryParser;
 import de.fraunhofer.iosb.ilt.frostserver.path.PathElementEntitySet;
@@ -81,11 +82,12 @@ public class CheckEntityQuery implements ValidationCheck {
     private void init(Entity contextEntity, JooqPersistenceManager pm) {
         entityType = contextEntity.getEntityType();
         final CoreSettings coreSettings = pm.getCoreSettings();
-        final QueryDefaults queryDefaults = coreSettings.getQueryDefaults();
-        final ResourcePath path = new ResourcePath(queryDefaults.getServiceRootUrl(), Version.V_1_1, '/' + entityType.plural)
+        final ResourcePath path = new ResourcePath("", Version.V_1_1, '/' + entityType.plural)
                 .addPathElement(new PathElementEntitySet(entityType));
         context = new DynamicContext();
-        parsedQuery = QueryParser.parseQuery(getQuery(), coreSettings, path, PrincipalExtended.INTERNAL_ADMIN_PRINCIPAL, context)
+        final QueryDefaults queryDefaults = coreSettings.getQueryDefaults();
+        final ModelRegistry modelRegistry = coreSettings.getModelRegistry();
+        parsedQuery = QueryParser.parseQuery(getQuery(), queryDefaults, modelRegistry, path, PrincipalExtended.INTERNAL_ADMIN_PRINCIPAL, context)
                 .validate(null, entityType);
         LOGGER.info("Initialised check on {}", entityType);
     }